feat/funnel 적용
