+++ src/app/signup/_components/CheckboxInput.tsx
@@ -3,7 +3,7 @@ const CheckboxInput = ({ label, blueLabel }: { label: string; blueLabel?: string
         <div className="flex place-items-center gap-[11px]">
             <input
                 type="checkbox"
-                className="appearance-none h-5 w-5 border-[1px] border-beigeLight rounded-[3px] bg-white checked:bg-primary checked:bg-[url('../assets/images/check.svg')] bg-no-repeat bg-center shadow-none"
+                className="appearance-none h-5 w-5 border-[1px] border-beigeLight rounded-[3px] bg-white checked:bg-primary checked:bg-[url('../assets/images/check.svg')] bg-no-repeat bg-[center_top_3px]"
             />
             <div className="text-[16px] font-semibold">
                 <span className="text-primary">{blueLabel}</span> {label}
