+++ src/app/signup/_components/LetterCount.tsx
